Add <requirement_level> to tool's reporting property.
<requirement_level> provides the relevance of the tool in regards to
the fulfillment of the associated criterion. It has three possible
values [REQUIRED, RECOMMENDED, OPTIONAL] in alignment with the
RFC 2119 requirement levels. Both [REQUIRED, RECOMMENDED] tools will
be executed by the QAA module, but only the outcome of the REQUIRED
will be taken into account for the criterion evaluation. The OPTIONAL
can be also requested to be executed by the QAA module.